% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.modreg.R
\name{methods}
\alias{methods}
\alias{coefficients.modreg}
\alias{coef.modreg}
\alias{print.modreg}
\alias{summary.modreg}
\alias{print.summary.modreg}
\title{Methods for fitted modreg model}
\usage{
\method{coefficients}{modreg}(object, ...)

\method{coef}{modreg}(object, ...)

\method{print}{modreg}(x, ...)

\method{summary}{modreg}(object, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}

\item{x, object}{A modreg object}
}
\value{
\code{coef} returns a named numerical vector with coefficients
}
\description{
Methods for \code{modreg} objects returned by the mode regression function.
}
